# src/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PortPulse AI API",
    "description": "PortPulse AI: forecasting, anomalies, CP-SAT optimisation, routing, and IBM Bob agent orchestration.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/scenarios": {
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Scenario",
        "operationId": "scenario_api_scenarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__routers__scenarios__ScenarioBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/extended": {
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Scenario Extended",
        "description": "W3 extended what-if: forecast and optimise the modified context, then compare.",
        "operationId": "scenario_extended_api_scenarios_extended_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendedScenarioBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioExtendedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/{scenario_id}/rollback": {
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Scenario Rollback",
        "operationId": "scenario_rollback_api_scenarios__scenario_id__rollback_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioRollbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/weather/refresh": {
      "post": {
        "tags": [
          "quality"
        ],
        "summary": "Weather Refresh",
        "description": "Run the W1 Open-Meteo pipeline and persist `WeatherObservation` rows.\n\nAudit B-6: the pipeline existed but was never invoked, so `FEATURE_WEATHER` was\nunreachable and `weather_used` stayed false forever. This is the missing wiring.\nIf Open-Meteo is unreachable the pipeline inserts nothing and this returns an\nempty series — the forecast then reports `weather_used: false` instead of failing.",
        "operationId": "weather_refresh_api_weather_refresh_post",
        "parameters": [
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 336,
              "minimum": 24,
              "default": 72,
              "title": "Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/weather": {
      "get": {
        "tags": [
          "quality"
        ],
        "summary": "Weather",
        "description": "Wind/wave/visibility series for San Pedro Bay (rows persisted by the pipeline).",
        "operationId": "weather_api_weather_get",
        "parameters": [
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 336,
              "minimum": 1,
              "default": 72,
              "title": "Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/bts/parse": {
      "post": {
        "tags": [
          "quality"
        ],
        "summary": "Bts Parse",
        "description": "Audit B-7: wire the BTS PPFSP benchmark parser to an endpoint (was dead code).\n\nAccepts a BTS weekly-berthing CSV, validates and parses it, returning the rows\nfor forecast validation. Keeps the benchmark load-bearing.",
        "operationId": "bts_parse_api_bts_parse_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_bts_parse_api_bts_parse_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_bts_parse_api_bts_parse_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_bts_parse_api_bts_parse_post"
      },
...
      "ExtendedScenarioBody": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "CRANE_OUTAGE | PRODUCTIVITY | BERTH_REMOVED | BERTH_ADDED | BUNCHING | SCHEDULE_CHANGE",
            "default": "CRANE_OUTAGE"
          },
          "crane_factor": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.5,
            "title": "Crane Factor",
            "default": 1.0
          },
          "move_rate_per_crane_hour": {
            "type": "number",
            "maximum": 35.0,
            "minimum": 20.0,
            "title": "Move Rate Per Crane Hour",
            "default": 28.0
          },
          "terminal_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal Code"
          },
          "berth_count_delta": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 0.0,
            "title": "Berth Count Delta",
            "default": 0
          },
          "bunching_vessels": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Bunching Vessels",
            "default": 0
          },
          "schedule_shift_hours": {
            "type": "number",
            "maximum": 48.0,
            "minimum": -48.0,
            "title": "Schedule Shift Hours",
            "default": -6.0
          },
          "parent_scenario_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Scenario Id"
          }
        },
        "type": "object",
        "title": "ExtendedScenarioBody",
        "description": "W3: berth add/remove, crane outage, vessel bunching, schedule change."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScenarioExtendedResponse": {
        "properties": {
          "baseline": {
            "additionalProperties": true,
            "type": "object",
            "title": "Baseline"
          },
          "scenario": {
            "additionalProperties": true,
            "type": "object",
            "title": "Scenario"
          },
          "impact": {
            "additionalProperties": true,
            "type": "object",
            "title": "Impact"
          },
          "feasible": {
            "type": "boolean",
            "title": "Feasible"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "parent_scenario_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Scenario Id"
          },
          "scenario_id": {
            "type": "integer",
            "title": "Scenario Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "forecast": {
            "additionalProperties": true,
            "type": "object",
            "title": "Forecast"
          },
          "solver": {
            "additionalProperties": true,
            "type": "object",
            "title": "Solver"
          },
          "weights": {
            "additionalProperties": true,
            "type": "object",
            "title": "Weights"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "baseline",
          "scenario",
          "impact",
          "feasible",
          "kind",
          "scenario_id",
          "description",
          "forecast",
          "solver",
          "weights"
        ],
        "title": "ScenarioExtendedResponse"
      },
      "ScenarioResponse": {
        "properties": {
          "scenario_id": {
            "type": "integer",
            "title": "Scenario Id"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "baseline": {
            "additionalProperties": true,
            "type": "object",
            "title": "Baseline"
          },
          "scenario": {
            "additionalProperties": true,
            "type": "object",
            "title": "Scenario"
          },
          "impact": {
            "additionalProperties": true,
            "type": "object",
            "title": "Impact"
          },
          "weights": {
            "additionalProperties": true,
            "type": "object",
            "title": "Weights"
          },
          "solver": {
            "additionalProperties": true,
            "type": "object",
            "title": "Solver"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scenario_id",
          "params",
          "baseline",
          "scenario",
          "impact",
          "weights",
          "solver"
        ],
        "title": "ScenarioResponse"
      },
      "ScenarioRollbackResponse": {
        "properties": {
          "restored": {
            "type": "boolean",
            "title": "Restored"
          },
          "scenario_id": {
            "type": "integer",
            "title": "Scenario Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "restored",
          "scenario_id",
          "status"
        ],
        "title": "ScenarioRollbackResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WeatherResponse": {
        "properties": {
          "points": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Points"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "hours": {
            "type": "integer",
            "title": "Hours"
          },
          "stub": {
            "type": "boolean",
            "title": "Stub",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "points",
          "source",
          "hours"
        ],
        "title": "WeatherResponse"
      },
...
      "app__routers__scenarios__ScenarioBody": {
        "properties": {
          "crane_factor": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.5,
            "title": "Crane Factor",
            "default": 1.0
          },
          "move_rate_per_crane_hour": {
            "type": "number",
            "maximum": 35.0,
            "minimum": 20.0,
            "title": "Move Rate Per Crane Hour",
            "default": 28.0
          }
        },
        "type": "object",
        "title": "ScenarioBody"
      }
    }
  }
}